More UI WIP

+++ SRC/UserMenu/UserFile.h
@@ -16,7 +16,7 @@ namespace UI
 
     size_t FileSize(const TCHAR * filename);
     void* FileLoad(const TCHAR * filename, size_t* size = nullptr);           // free!
-    bool FileSave(const TCHAR * filename, void* data, uint32_t size);
+    bool FileSave(const TCHAR * filename, void* data, size_t size);
     TCHAR* FileOpen(HWND hwnd, FileType type = FileType::All);        // copy away!
     TCHAR* FileShortName(const TCHAR * filename, int lvl = 3);           // copy away!
     TCHAR* FileSmartSize(size_t size);                            // copy away!
